fix(geo,admin): cast cities.id to int in SELECTs, not raw bigint string
pg returns BIGINT columns as JS strings (no type-parser override for OID
20), but every consumer types City.id as number. The driver onboarding
city dropdown picked up that string id and sent it as city_id in the
personal-info payload, which the zod schema (z.number()) rejected with
"Expected number, received string".

+++ api/src/modules/admin/admin.repository.ts
@@ -882,7 +882,7 @@ export async function listExpiringDocs(daysAhead: number): Promise<ExpiringVehic
 // ─── Geo / Cities ─────────────────────────────────────────────────────────────
 
 const ADMIN_CITY_COLS = `
-  id, name, slug, state,
+  id::int, name, slug, state,
   ST_Y(centroid::geometry) AS centroid_lat,
   ST_X(centroid::geometry) AS centroid_lng,
   default_speed_limit_kmph,
